+++ packages/wallets/src/evm/connectors/embedded-wallet/index.ts
@@ -237,7 +237,7 @@ export class EmbeddedWalletConnector extends Connector<EmbeddedWalletConnectionA
         });
       }
       case "apple":
-      // case "facebook":
+      case "facebook":
       case "google": {
         const oauthProvider = oauthStrategyToAuthProvider[strategy];
         return ewSDK.auth.loginWithOauth({
@@ -276,6 +276,6 @@ const oauthStrategyToAuthProvider: Record<
   AuthProvider
 > = {
   google: AuthProvider.GOOGLE,
-  // facebook: AuthProvider.FACEBOOK,
+  facebook: AuthProvider.FACEBOOK,
   apple: AuthProvider.APPLE,
 };